cambios generales y de vista

+++ app/Database/Migrations/2026-06-01-000003_sgc_casos_coordenadas.php
@@ -26,11 +26,11 @@ class CasosCoordenadas extends Migration
             ],
             'latitud' => [
                 'null' => true,
-                'type' => 'DOUBLE',
+                'type' => 'FLOAT',
             ],
             'longitud' => [
                 'null' => true,
-                'type' => 'DOUBLE',
+                'type' => 'FLOAT',
             ],
             'fecha_creacion' => [
                 'null' => true,
